# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ДоВуз-Портал API
  version: 0.1.0
  description: |
    Контракт бэкенда АИС «ДоВуз-Портал». Снят со слоя `frontend/src/app/api/`
    и повторяет его как есть, включая смешанное именование полей: часть ответов
    в snake_case, часть в camelCase, query-параметры в camelCase.

    Аутентификация — HTTP-only кука сессии, которую выставляют `/auth/login`
    и `/auth/register`. Ошибки всегда приходят в формате `{"message": "..."}`.
    401 — «не вошёл» (фронт разлогинивает), 403 — «вошёл, но нельзя».
servers:
  - url: /v1
    description: Относительно хоста, с которого отдана спека; локально — http://localhost:8080/v1
security:
  - sessionCookie: []
...
paths:
...
  /subjects:
    get:
      tags: [subjects]
      operationId: getSubjects
      summary: Предметы направления
      parameters:
        - name: directionId
          in: query
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница предметов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [subjects]
      operationId: createSubject
      summary: Создать предмет
      description: Только для администратора. Название уникально в пределах направления.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreateRequest'
      responses:
        '201':
          description: Предмет создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /subjects/{id}:
    get:
      tags: [subjects]
      operationId: getSubject
      summary: Предмет по id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Предмет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [subjects]
      operationId: updateSubject
      summary: Переименовать предмет
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        '200':
          description: Предмет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [subjects]
      operationId: deleteSubject
      summary: Удалить предмет
      description: Вместе с темами, публикациями и их файлами.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Удалён
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dovuz_session
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageNumber:
      name: pageNumber
      in: query
      description: Номер страницы, с нуля
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 10
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Нужно войти в систему
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Name:
      type: string
      minLength: 1
      maxLength: 300
    PageMeta:
      type: object
      required: [total_pages, total_size, page_number, page_size]
      properties:
        total_pages:
          type: integer
        total_size:
          type: integer
        page_number:
          type: integer
        page_size:
          type: integer
...
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          $ref: '#/components/schemas/Name'
...
    SubjectResponse:
      type: object
      required: [id, name, directionId, directionName]
      properties:
        id:
          type: integer
        name:
          type: string
        directionId:
          type: integer
        directionName:
          type: string
    SubjectCreateRequest:
      type: object
      required: [name, directionId]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        directionId:
          type: integer
    SubjectPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/SubjectResponse'
...